Fix Code Standard
