+++ src/main/java/duke/commands/DeleteCommand.java
@@ -36,8 +36,6 @@ public class DeleteCommand extends Command {
     public String execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         Task task = tasks.getTask(this.index - 1);
         tasks.delete(this.index - 1);
-//        ui.showLine();
-//        ui.showLine();
         storage.writeData(tasks.getAllTasks());
         return ui.showDelete(this.index, task);
     }
