+++ src/main/java/duke/commands/FindCommand.java
@@ -31,9 +31,7 @@ public class FindCommand extends Command {
      */
     @Override
     public String execute(TaskList tasks, Ui ui, Storage storage) {
-//        ui.showLine();
         return ui.showFind(this.find) + "\n" + tasks.filterByName(this.find);
-//        ui.showLine();
     }
 
     /**
