+++ src/main/java/duke/commands/ListCommand.java
@@ -20,10 +20,6 @@ public class ListCommand extends Command {
      */
     @Override
     public String execute(TaskList tasks, Ui ui, Storage storage) {
-//        ui.showLine();
-//        ;
-//        tasks.print();
-//        ui.showLine();
         return ui.showList(tasks.size()) + "\n" + tasks.toString();
     }
