+++ src/main/java/duke/tasks/TaskList.java
@@ -106,7 +106,7 @@ public class TaskList {
     public String toString() {
         StringBuilder taskString = new StringBuilder();
         for (int i = 0; i < this.tasks.size(); i++) {
-            taskString.append((i + 1) + ") " + tasks.get(i).toString() + "\n");
+            taskString.append(i + 1).append(") ").append(tasks.get(i).toString()).append("\n");
         }
         return taskString.toString();
     }
