+++ src/main/resources/css/Application.css
@@ -5,7 +5,14 @@
 }
 
 .display-picture {
-    -fx-border-radius: 20;
+    -fx-border-radius: 50px;
+    -fx-hgap: 20px;
+    -fx-padding: 40px;
+    -fx-background-color: #2969c0;
+    -fx-background-radius: 50px;
+    -fx-border-width: 5px;
+    -fx-border-color: black;
+    -fx-effect: dropshadow(three-pass-box, rgba(0, 0, 0, 0.8), 10, 0, 0, 0);
 }
 
 .message-text {
